requirements.txt
PyMuPDF
python-docx
matplotlib
pandas
pillow
numpy
webbrowser-open